[fix] Fine tune paramater type

+++ srcs/ft_flags_pad_utils_bonus.c
@@ -6,13 +6,13 @@
 /*   By: lyeh <lyeh@student.42vienna.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/19 11:09:13 by lyeh              #+#    #+#             */
-/*   Updated: 2023/09/21 18:03:26 by lyeh             ###   ########.fr       */
+/*   Updated: 2023/09/22 21:25:58 by lyeh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf_bonus.h"
 
-char	*ft_pad_on_right(char *str, char pad, char type, size_t total_len)
+char	*ft_pad_on_right(char *str, char pad, char type, int total_len)
 {
 	char	*ret;
 	int		len;
@@ -41,7 +41,7 @@ char	*ft_pad_on_right(char *str, char pad, char type, size_t total_len)
 	return (ret);
 }
 
-char	*ft_pad_on_left(char *str, char pad, char type, size_t total_len)
+char	*ft_pad_on_left(char *str, char pad, char type, int total_len)
 {
 	char	*ret;
 	int		len;
