+++ srcs/ft_format_string_bonus.c
@@ -6,7 +6,7 @@
 /*   By: lyeh <lyeh@student.42vienna.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/15 16:01:22 by lyeh              #+#    #+#             */
-/*   Updated: 2023/09/21 18:03:55 by lyeh             ###   ########.fr       */
+/*   Updated: 2023/09/22 21:25:29 by lyeh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ char	*ft_format_string(char *s, t_print_tab *tab)
 		s = "(null)";
 	if (ft_strncmp(s, "(null)", ft_strlen(s)) == 0 && tab->f_perc_fmt)
 	{
-		if (tab->perc_len >= ft_strlen(s))
+		if (tab->perc_len >= (int)ft_strlen(s))
 			tmp = ft_strdup(s);
 		else
 			tmp = ft_strdup("");
